Add messageboxes to MainViewModel

+++ JustNote/Viewmodels/MainViewModel.cs
@@ -1,5 +1,7 @@
 ﻿
-// Author: Lukáš Leták
+// Authors:
+//  Nikola Machálková – message boxes
+//        Lukáš Leták – the rest
 
 using System;
 using System.Collections.Generic;
@@ -14,6 +16,7 @@ using JustNote.App.Services;
 using JustNote.Backend.Data;
 using Microsoft.Toolkit.Mvvm.Input;
 using System.Diagnostics;
+using System.Windows;
 
 namespace JustNote.App.Viewmodels
 {
@@ -26,7 +29,7 @@ namespace JustNote.App.Viewmodels
         private DateTime _Date;
         private string _Title;
         private int _noteID;
-     
+
 
         public MainViewModel(IDataService dataService)
         {
@@ -167,20 +170,30 @@ namespace JustNote.App.Viewmodels
 
         private void RemoveTextbox(int key)
         {
-            foreach (var note in Notes)
+            var result = MessageBox.Show("Are you sure you want to delete note? This can't be undone.", "Warning", System.Windows.MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (result == MessageBoxResult.Yes)
             {
-                if (note.Key == key)
+                foreach (var note in Notes)
                 {
-                    Notes.Remove(note);
-                    return;
+                    if (note.Key == key)
+                    {
+                        Notes.Remove(note);
+                        return;
+                    }
                 }
+                MessageBox.Show("Action was done successfully.", "Notice", MessageBoxButton.OK, MessageBoxImage.Information);
             }
         }
 
         private void DeleteWholeNote()
         {
-            _dataService.DeleteDateData(DateData);
-            LoadDateData(DateData.Date, null);
+            var result = MessageBox.Show("Are you sure you want to delete page? This can't be undone.", "Warning", System.Windows.MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (result == MessageBoxResult.Yes)
+            {
+                _dataService.DeleteDateData(DateData);
+                LoadDateData(DateData.Date, null);
+                MessageBox.Show("Action was done successfully.", "Notice", MessageBoxButton.OK, MessageBoxImage.Information);
+            }
         }
 
         private void LoadDateData(DateTime date, string? title)
@@ -203,7 +216,6 @@ namespace JustNote.App.Viewmodels
 
         private void SaveDateData()
         {
-            Debug.WriteLine("saving");
             DateData.Date = Date;
             DateData.Title = Title;
             if (DateData.Notes == null)
@@ -215,6 +227,8 @@ namespace JustNote.App.Viewmodels
                 DateData.Notes.Add(normalNote);
             }
             _dataService.SaveDateData(DateData);
+
+            MessageBox.Show("Note was saved succesfully.", "Notice", System.Windows.MessageBoxButton.OK, MessageBoxImage.Information);
         }
     }
 }
